refactor: split session create handlers
